atualização de rotas

+++ src/main/java/com/marco/sistemaVendas/services/CategoriaServices.java
@@ -56,6 +56,9 @@ public class CategoriaServices {
 	public void delete(Long id) {
 		if(repositories.existsById(id)) {
 			try {
+				if(repositories.findById(id).get().getProdutos().size() > 0) {
+					throw new DatabaseException("Cannot delete or update a parent row");
+				}
 				repositories.deleteById(id);
 			} catch (DataIntegrityViolationException e) {
 				throw new DatabaseException(e.getMessage());
