+++ src/main/java/com/marco/sistemaVendas/services/ProdutoServices.java
@@ -53,6 +53,9 @@ public class ProdutoServices {
 	public void delete(Long id) {
 		if(repositories.existsById(id)) {
 			try {
+				if(repositories.findById(id).get().getProdutoDaVenda().size() > 0) {
+					throw new DatabaseException("Cannot delete or update a parent row");
+				}
 				repositories.deleteById(id);
 			} catch (DataIntegrityViolationException e) {
 				throw new DatabaseException(e.getMessage());
